lab3

+++ LabB/src/main/java/mk/ukim/finki/wp/lab/service/impl/TeacherServiceImpl.java
@@ -1,6 +1,8 @@
 package mk.ukim.finki.wp.lab.service.impl;
 
 import mk.ukim.finki.wp.lab.model.Teacher;
+import mk.ukim.finki.wp.lab.model.exception.TeacherAlreadyExists;
+import mk.ukim.finki.wp.lab.model.exception.TeacherNotFoundedException;
 import mk.ukim.finki.wp.lab.repository.jpa.CoursesRepositoryJpa;
 import mk.ukim.finki.wp.lab.repository.jpa.TeacherRepositoryJpa;
 import mk.ukim.finki.wp.lab.service.TeacherService;
@@ -27,27 +29,31 @@ public class TeacherServiceImpl implements TeacherService {
 
     @Override
     public Optional<Teacher> findById(Long id) {
-        return this.teacherRepository.findById(id);
+        return Optional.of(this.teacherRepository.findById(id).orElseThrow(()-> new TeacherNotFoundedException(id)));
     }
 
     @Override
     public Teacher save(String name, String surname) {
             if(this.teacherRepository.findAll().stream().noneMatch(t->t.getName().equals(name))){
-                return new Teacher(name,surname);
-            }
-        return null;
+                return this.teacherRepository.save(new Teacher(name,surname));
+            }else throw new TeacherAlreadyExists(name);
     }
 
     @Override
-    public Teacher addCourse(long teacherId, long courseId) {
-        this.courseRepository.findById(courseId).get().setTeacher(this.teacherRepository.findById(teacherId));
-        this.teacherRepository.findById(teacherId).getCourses()
-                .add(this.courseRepository.findById(courseId).get());
+    public Optional<Teacher> addCourse(long teacherId, long courseId) {
+        if(this.teacherRepository.findById(teacherId).isPresent()){
+            if(this.courseRepository.findById(courseId).isPresent()){
+                this.teacherRepository.findById(teacherId).get().getCourses()
+                        .add(this.courseRepository.findById(courseId).get());
+            }
+        }
         return this.teacherRepository.findById(teacherId);
     }
 
     @Override
     public void deleteById(long teacherId) {
-        this.teacherRepository.deleteById(teacherId);
+        if(this.teacherRepository.findById(teacherId).isPresent()) {
+            this.teacherRepository.deleteById(teacherId);
+        }else throw new TeacherNotFoundedException(teacherId);
     }
 }
